настроена работа с базой данных, удаление и добавление вакансии и избранное, поиск деталей вакансии через hh api, создан mapper для entity и dto моделей

+++ app/src/main/java/ru/practicum/android/diploma/favorites/presentation/FavoritesViewModel.kt
@@ -60,6 +60,6 @@ class FavoritesViewModel(private val favoritesInteractor: FavoritesInteractor) :
     }
 
     private fun onClickDebounce(item: VacancyShort) {
-        screenStateLiveData.postValue(ShortVacancyListUiState.AnyItem(item.vacancyId))
+        screenStateLiveData.postValue(ShortVacancyListUiState.AnyItem(item.vacancyId.toInt()))
     }
 }
